Removed decorator, moved cache in to main Github service
Because the API might throw rate limits, the service needs to know about the cache to have a fallback.
