+++ ReadingList.Notes.Github/Repositories/IGithubTextFileService.cs
@@ -1,12 +1,14 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Octokit;
 using Readinglist.Notes.Logic.Models;
 
 namespace ReadingList.Notes.Github.Repositories
 {
-    public interface IGithubTextFileService
+    public interface IGithubBookRecordService
     {
+        Task<IEnumerable<BookRecord>> GetAllBookRecords();
         Task<BookRecord> GetBookRecord(RepositoryContent content);
     }
 }
